project/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
djangorestframework==3.11.1
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
